+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Add
 
 - `set` method on `collSpecable`;
+- Associate a random id on array items stores by default, instead of using indexes (which varies too much when dealing with reordering, deletions, etc.);
 
 ### Fix
 
 - Ensure promise is always returned from `activate` method;
+- Array store values not properly saved after removal;
 
 ---